Linked View Controllers
Allowed the View Controllers the ability to interact with eachother within a prepare function in the main View Controller.

Also Connected the outlet information to the Task object which assigned to a constant of task.

Lastly gave the addTapped function the ability to reload the tableView in the previous View Controller with the new task data and revert the screen back to the main View Controller.